+++ projects/cloud-airlines-angular-sdk/src/components/abstract.component.ts
@@ -4,6 +4,8 @@ import { TranslateService } from '@ngx-translate/core';
 import { I18n } from "../i18n/i18n.interface";
 import { DateAdapter } from "@angular/material/core";
 import { CfarOffer, MapString } from "../apis/hopper-cloud-airline/v1";
+import { CountryCode } from "../enums/country-code.enum";
+import { take } from "rxjs/operators";
 
 @Directive({
     selector: '[HopperAbstractComponent]'
@@ -14,6 +16,8 @@ export abstract class AbstractComponent implements OnChanges {
     @Input() basePath!: string;
     @Input() isFakeBackend!: boolean;
     @Input() extAttributes!: MapString;
+
+    public mapCountries: Map<string, string>;
 
     constructor(
         protected adapter: DateAdapter<any>,
@@ -32,6 +36,9 @@ export abstract class AbstractComponent implements OnChanges {
 
         // Set default language for datepickers
         this.adapter.setLocale(this.translateService.getBrowserLang());
+
+        // Init map
+        this.mapCountries = new Map<string, string>();
     }
     
     // -----------------------------------------------
@@ -40,8 +47,11 @@ export abstract class AbstractComponent implements OnChanges {
 
     ngOnChanges(changes: SimpleChanges): void {
         if (changes.currentLang && changes.currentLang.currentValue) {
-            // update current language
-            this.translateService.use(changes.currentLang.currentValue);
+            // Update languages
+            this._updateLanguage(changes.currentLang.currentValue);
+
+            // Update countries labels
+            this._setCountriesLabels();
         }
     }
 
@@ -53,5 +63,27 @@ export abstract class AbstractComponent implements OnChanges {
         return offers.reduce((previous: CfarOffer, current: CfarOffer) => {
             return +previous?.premium < +current?.premium ? previous : current;
         });
+    }
+
+    protected _updateLanguage(newLanguage: string): void {
+        // Set language for components
+        this.translateService.use(newLanguage);
+
+        // Set language for datepickers
+        this.adapter.setLocale(newLanguage);
+    }
+
+    protected _setCountriesLabels(): void {
+        const countries = Object.keys(CountryCode);
+
+        countries.forEach(countryCode => {
+            // Get Label and fill the map
+            this.translateService.get('COMMON.COUNTRY.' + countryCode)
+                .pipe(take(1))
+                .subscribe(label => this.mapCountries.set(countryCode, label));            
+        });
+        
+        // Sort the map by label (alphabetical order)
+        this.mapCountries = new Map([...this.mapCountries.entries()].sort((a, b) => a[1].localeCompare(b[1])));
     }
 }